requirements.txt
transformers==4.34.1
tokenizers==0.14.1
huggingface-hub==0.17.3
accelerate==0.25.0
diffusers==0.23.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
hf-xet==1.2.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
jsonpointer==2.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==12.0.0
protobuf==3.20.3
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
setuptools==78.1.1
sympy==1.14.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
